backend/requirements.txt
fastapi
uvicorn
pillow
torch
open_clip
transformers
